+++ app/services/transaction_service.py
@@ -3,7 +3,11 @@ from app.schemas.transaction import TransactionCreate, TransactionUpdate
 from app.repositories.account_repository import AccountRepository
 from app.services.fraud_service import FraudDetectionService
 from uuid import UUID
-from app.exceptions import InsufficientBalanceError, TransactionAlreadyProcessedError
+from app.exceptions import (
+    InsufficientBalanceError,
+    TransactionAlreadyProcessedError,
+    TransactionAccessDeniedError
+)
 from app.models.transaction import TransactionStatus
 
 class TransactionService:
@@ -14,13 +18,19 @@ class TransactionService:
         self.account_repository = account_repository
         self.fraud_service = fraud_service
 
-    def create_transaction(self, transaction: TransactionCreate):
+    def create_transaction(self, transaction: TransactionCreate, user_id: UUID):
         account = self.account_repository.get_by_id(
             transaction.account_id
         )
 
         if account is None:
             return None
+
+        if account.user_id != user_id:
+            raise TransactionAccessDeniedError(
+                "You do not have access to this account"
+            )
+
         
         if transaction.amount > account.balance:
             raise InsufficientBalanceError(
@@ -75,4 +85,91 @@ class TransactionService:
         )
 
     def delete_transaction(self, transaction_id):
+        return self.repository.delete(transaction_id)
+
+    def get_transaction_for_user(self, transaction_id: UUID, user_id: UUID):
+        transaction = self.repository.get_by_id(transaction_id)
+
+        if transaction is None:
+            return None
+
+        account = self.account_repository.get_by_id(
+            transaction.account_id
+        )
+
+        if account is None:
+            return None
+
+        if account.user_id != user_id:
+            raise TransactionAccessDeniedError(
+                "You do not have access to this transaction"
+            )
+
+        return transaction
+
+    def get_transactions_for_user(self, user_id: UUID):
+        accounts = self.account_repository.get_by_user_id(user_id)
+
+        if not accounts:
+            return []
+
+        account_ids = [account.id for account in accounts]
+
+        return self.repository.get_by_account_ids(account_ids)
+
+    def update_transaction_for_user(
+        self,
+        transaction_id: UUID,
+        transaction: TransactionUpdate,
+        user_id: UUID
+    ):
+        db_transaction = self.repository.get_by_id(transaction_id)
+
+        if db_transaction is None:
+            return None
+
+        account = self.account_repository.get_by_id(
+            db_transaction.account_id
+        )
+
+        if account is None:
+            return None
+
+        if account.user_id != user_id:
+            raise TransactionAccessDeniedError(
+                "You do not have access to this transaction"
+            )
+
+        if db_transaction.status != TransactionStatus.PENDING:
+            raise TransactionAlreadyProcessedError(
+                "Processed transactions cannot be modified"
+            )
+
+        return self.repository.update(
+            transaction_id,
+            transaction
+        )
+
+    def delete_transaction_for_user(
+        self,
+        transaction_id: UUID,
+        user_id: UUID
+    ):
+        db_transaction = self.repository.get_by_id(transaction_id)
+
+        if db_transaction is None:
+            return None
+
+        account = self.account_repository.get_by_id(
+            db_transaction.account_id
+        )
+
+        if account is None:
+            return None
+
+        if account.user_id != user_id:
+            raise TransactionAccessDeniedError(
+                "You do not have access to this transaction"
+            )
+
         return self.repository.delete(transaction_id)
